+++ src/main/java/cn/edu/seu/kse/lpmln/translator/ASPTranslatorV2.java
@@ -52,7 +52,7 @@ public class ASPTranslatorV2 extends  ASPTranslator{
     @Override
     protected String translateCountingPart(Rule rule, boolean isSoft){
         return new StringBuilder().append(":~").append(satLabel).append(".")
-                .append(" [").append(isSoft?((int)(rule.getWeight()*factor)+"@1, "):"1@2, ")
+                .append(" [").append(isSoft?((long)(rule.getWeight()*factor)+"@1, "):"1@2, ")
                 .append(rule.getId()).append(rule.getVars().size()>0?", ":"").append(String.join(",",rule.getVars())).append("]")
                 .append(System.lineSeparator()).toString();
     }
