+++ src/main/java/cn/edu/seu/kse/lpmln/translator/DLVTranslator.java
@@ -18,7 +18,7 @@ public class DLVTranslator extends ASPTranslator {
         sb.append(":~ sat(").append(rule.getRuleLabel()).append("). ");
         sb.append("[");
         if(isSoft){
-            int weight= (int) (rule.getWeight()*factor);
+            long weight= (long) (rule.getWeight()*factor);
             sb.append(weight).append(":1]");
         }else {
             sb.append("1:2] ");
